Fix get_integer() warnings - strtol returns a long, not an int...
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@7360 7a7537e8-13f0-0310-91df-b6672ffda945

+++ ppdc/ppdc-source.cxx
@@ -1183,8 +1183,8 @@ ppdcSource::get_installable(ppdcFile *fp)
 int					// O - Integer value
 ppdcSource::get_integer(const char *v)	// I - Value string
 {
-  int	val;				// Value
-  int	temp;				// Temporary value
+  long	val;				// Value
+  long	temp;				// Temporary value
   char	*newv;				// New value string pointer
 
 
@@ -1199,7 +1199,7 @@ ppdcSource::get_integer(const char *v)	// I - Value string
     if (*v || val == LONG_MIN)
       return (-1);
     else
-      return (val);
+      return ((int)val);
   }
   else if (*v == '(')
   {
@@ -1220,7 +1220,7 @@ ppdcSource::get_integer(const char *v)	// I - Value string
     }
 
     if (*v == ')')
-      return (val);
+      return ((int)val);
     else
       return (-1);
   }
